# parser/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dmarc-parser API",
    "version": "2.0.0",
    "description": "REST API of the DMARC aggregate report parser & ingestion service.\n\n### Authentication\nEvery `/api/v1/*` route (except `/api/v1/openapi.json`) requires an API\nkey from `PARSER_API_KEYS`, sent either as `Authorization: Bearer <key>`\nor `X-API-Key: <key>`. Keys carry scopes from `{read, ingest, admin}`;\neach operation documents its required scope via `x-required-scope`.\nMissing/invalid key \u2192 401; key without the required scope \u2192 403.\n\n### Rate limiting\nPer-key token bucket (`PARSER_RATE_LIMIT_RPS`, default 10 req/s,\nburst 3x). Exceeding it returns 429 with a `Retry-After` header.\n\n### Conventions\nAll timestamps are UTC, RFC 3339. Date-only values accepted by `since` /\n`until` parameters use `YYYY-MM-DD`. Errors are JSON\n`{\"error\": \"message\"}`.\n"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Host loopback"
    },
    {
      "url": "http://dmarcparser:8080",
      "description": "From containers on the mxsentinel_default / dmarcparser_default networks"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/domains": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Domain overview with 30d volume, alignment, policy and health score",
        "operationId": "listDomains",
        "x-required-scope": "read",
        "responses": {
          "200": {
            "description": "Domain list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "domains": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DomainOverview"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/v1/domains/{domain}/readiness": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Policy enforcement readiness verdict",
        "description": "Pure rule engine over 30/90-day alignment and failing sources. Recommendation is one of enforce | step_pct | fix_alignment | monitor.\n",
        "operationId": "getDomainReadiness",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/domainPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Readiness verdict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/domains/{domain}/selectors": {
      "get": {
        "tags": [
          "reports2"
        ],
        "summary": "DKIM selector inventory for one domain",
        "operationId": "getDomainSelectors",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/domainPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Selector inventory with first/last seen and pass/fail counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "domain": {
                      "type": "string"
                    },
                    "selectors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Selector"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/v1/stats/timeseries": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Message volume / alignment timeseries from daily rollups",
        "operationId": "getStatsTimeseries",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/domain"
          },
          {
            "name": "bucket",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week"
              ],
              "default": "day"
            }
          },
          {
            "$ref": "#/components/parameters/since"
          },
          {
            "$ref": "#/components/parameters/until"
          }
        ],
        "responses": {
          "200": {
            "description": "Timeseries points",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string",
                      "enum": [
                        "day",
                        "week"
                      ]
                    },
                    "points": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TimeseriesPoint"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/stats/top": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Top senders along one dimension",
        "operationId": "getStatsTop",
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "dimension",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ip",
                "org",
                "header_from"
              ],
              "default": "ip"
            }
          },
          {
            "$ref": "#/components/parameters/domain"
          },
          {
            "$ref": "#/components/parameters/since"
          },
          {
            "$ref": "#/components/parameters/until"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "failing",
            "in": "query",
            "description": "Restrict to fully unaligned records",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked rows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "dimension": {
                      "type": "string",
                      "enum": [
                        "ip",
                        "org",
                        "header_from"
                      ]
                    },
                    "failing": {
                      "type": "boolean"
                    },
                    "rows": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TopRow"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/tlsrpt": {
      "get": {
        "tags": [
          "reports2"
        ],
        "summary": "List SMTP TLS (RFC 8460) reports",
        "operationId": "listTLSRPT",
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "description": "Filter by policy domain",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/org"
          },
          {
            "$ref": "#/components/parameters/since"
          },
          {
            "$ref": "#/components/parameters/until"
          },
          {
            "$ref": "#/components/parameters/limit500"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "responses": {
          "200": {
            "description": "TLS-RPT report summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reports": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TLSRPTSummary"
                      }
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/tlsrpt/{id}": {
      "get": {
        "tags": [
          "reports2"
        ],
        "summary": "One TLS-RPT report with per-policy rows",
        "operationId": "getTLSRPT",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "TLS-RPT detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TLSRPTDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/forensic": {
      "get": {
        "tags": [
          "reports2"
        ],
        "summary": "List forensic (ruf) reports",
        "description": "Recipient localparts are redacted at store time when PARSER_RUF_REDACT is on.",
        "operationId": "listForensic",
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "description": "Filter by reported domain",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/since"
          },
          {
            "$ref": "#/components/parameters/until"
          },
          {
            "$ref": "#/components/parameters/limit500"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Forensic report list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reports": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ForensicReport"
                      }
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/forensic/{id}": {
      "get": {
        "tags": [
          "reports2"
        ],
        "summary": "One forensic report (includes raw headers)",
        "operationId": "getForensic",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Forensic report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForensicReport"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from PARSER_API_KEYS as a bearer token. Key scopes: read, ingest, admin (see each operation's x-required-scope)."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same API key via the X-API-Key header."
      }
    },
    "parameters": {
      "domain": {
        "name": "domain",
        "in": "query",
        "description": "Filter by policy-published domain",
        "schema": {
          "type": "string"
        }
      },
      "org": {
        "name": "org",
        "in": "query",
        "description": "Filter by reporting organisation",
        "schema": {
          "type": "string"
        }
      },
      "since": {
        "name": "since",
        "in": "query",
        "description": "Lower time bound (RFC 3339 or YYYY-MM-DD)",
        "schema": {
          "type": "string"
        }
      },
      "until": {
        "name": "until",
        "in": "query",
        "description": "Upper time bound (RFC 3339 or YYYY-MM-DD)",
        "schema": {
          "type": "string"
        }
      },
      "limit500": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 50
        }
      },
...
      "offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
...
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "domainPath": {
        "name": "domain",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "API key lacks the required scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Per-key rate limit exceeded (see Retry-After)",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "TimeseriesPoint": {
        "type": "object",
        "properties": {
          "bucket": {
            "type": "string",
            "description": "Bucket start date (YYYY-MM-DD)"
          },
          "msgs_total": {
            "type": "integer",
            "format": "int64"
          },
          "msgs_aligned": {
            "type": "integer",
            "format": "int64"
          },
          "pass_rate": {
            "type": "number"
          },
          "msgs_none": {
            "type": "integer",
            "format": "int64"
          },
          "msgs_quarantine": {
            "type": "integer",
            "format": "int64"
          },
          "msgs_reject": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TopRow": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Dimension value (ip",
            "org or header_from)": null
          },
          "msgs": {
            "type": "integer",
            "format": "int64"
          },
          "msgs_failed": {
            "type": "integer",
            "format": "int64"
          },
          "reports": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DomainOverview": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "msgs_30d": {
            "type": "integer",
            "format": "int64"
          },
          "aligned_rate_30d": {
            "type": "number"
          },
          "last_report": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "policy_p": {
            "type": [
              "string",
              "null"
            ]
          },
          "policy_sp": {
            "type": [
              "string",
              "null"
            ]
          },
          "policy_pct": {
            "type": [
              "integer",
              "null"
            ]
          },
          "health": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          }
        }
      },
...
      "Readiness": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "aligned_rate_30d": {
            "type": "number"
          },
          "aligned_rate_90d": {
            "type": "number"
          },
          "msgs_30d": {
            "type": "integer",
            "format": "int64"
          },
          "msgs_90d": {
            "type": "integer",
            "format": "int64"
          },
          "fail_sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FailSource"
            }
          },
          "current_policy": {
            "type": "object",
            "properties": {
              "p": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "sp": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "pct": {
                "type": [
                  "integer",
                  "null"
                ]
              }
            }
          },
          "recommendation": {
            "type": "string",
            "enum": [
              "enforce",
              "step_pct",
              "fix_alignment",
              "monitor"
            ]
          },
          "blockers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Selector": {
        "type": "object",
        "properties": {
          "dkim_domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "selector": {
            "type": [
              "string",
              "null"
            ]
          },
          "first_seen": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "last_seen": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "pass_count": {
            "type": "integer",
            "format": "int64"
          },
          "fail_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TLSRPTSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "org": {
            "type": "string"
          },
          "report_id": {
            "type": "string"
          },
          "contact": {
            "type": [
              "string",
              "null"
            ]
          },
          "date_begin": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "date_end": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "seen": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "policies": {
            "type": "integer",
            "format": "int64"
          },
          "success_total": {
            "type": "integer",
            "format": "int64"
          },
          "failure_total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "TLSRPTDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TLSRPTSummary"
          },
          {
            "type": "object",
            "properties": {
              "policy_rows": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TLSRPTPolicyRow"
                }
              }
            }
          }
        ]
      },
      "ForensicReport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "seen": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "feedback_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "auth_failure": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_ip": {
            "type": [
              "string",
              "null"
            ]
          },
          "reported_domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "original_mail_from": {
            "type": [
              "string",
              "null"
            ]
          },
          "arrival_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "subject": {
            "type": [
              "string",
              "null"
            ]
          },
          "message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "header_from": {
            "type": [
              "string",
              "null"
            ]
          },
          "raw_headers": {
            "type": [
              "string",
              "null"
            ],
            "description": "Only present on the single-report endpoint"
          }
        }
      },
...
    }
  }
}